refactored get all notes route to method on user schema
